Completed topological dependency loading function and moved ModUtils as can be used by non-cs packages.
