+++ Barotrauma/BarotraumaShared/SharedSource/LuaCs/Plugins/CsScriptManager.cs
@@ -79,7 +79,9 @@ public class CsScriptManager
     {
         var syntaxTree = new List<SyntaxTree>();
         syntaxTree.AddRange(GeneratePackageScriptAssemblyInfo(package));
-        
+
+
+        throw new NotImplementedException();
     }
 
     public static SyntaxTree GetPackageScriptImports() => AssemblyImports;
@@ -109,97 +111,6 @@ public class CsScriptManager
     {
         throw new NotImplementedException();
     }
-    
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="packages">A dictionary/map with key as the package and the elements as it's dependencies.</param>
-    /// <param name="readyToLoad"></param>
-    /// <param name="cannotLoadPackages">Packages with errors or cyclic dependencies.</param>
-    /// <returns>Whether or not the process produces a usable list.</returns>
-    private static bool OrderAndFilterPackagesByDependencies(
-        Dictionary<ContentPackage, IEnumerable<ContentPackage>> packages, 
-        out IEnumerable<ContentPackage> readyToLoad, 
-        out IEnumerable<KeyValuePair<ContentPackage, string>> cannotLoadPackages)
-    {
-        HashSet<ContentPackage> completedPackages = new();
-        List<ContentPackage> readyPackages = new();
-        Dictionary<ContentPackage, string> unableToLoad = new();
-        HashSet<ContentPackage> currentNodeChain = new();
-
-        readyToLoad = readyPackages;
-        cannotLoadPackages = unableToLoad;
-
-        foreach (var toProcessPack in packages)
-        {   
-            ProcessPackage(toProcessPack.Key, toProcessPack.Value);
-        }
-
-        PackageProcRet ProcessPackage(ContentPackage packageToProcess, IEnumerable<ContentPackage> dependencies)
-        {
-            // already processed
-            if (completedPackages.Contains(packageToProcess))
-            {
-                return PackageProcRet.AlreadyCompleted;
-            }
-            
-            // cyclic check
-            if (currentNodeChain.Contains(packageToProcess))
-            {
-                StringBuilder sb = new();
-                sb.AppendLine("Error: Cyclic Dependency. ")
-                    .Append("The following ContentPackages rely on eachother in a way that makes it impossible to know which to load first! ")
-                    .Append("Note: the package listed twice shows where the cycle starts/ends and is not necessarily the problematic package.");
-                int i = 0;
-                foreach (var package in currentNodeChain)
-                {
-                    i++;
-                    sb.AppendLine($"{i}. {package.Name}");
-                }
-                sb.AppendLine($"{i}. {packageToProcess.Name}");
-                unableToLoad.Add(packageToProcess, sb.ToString());
-                return PackageProcRet.Cyclic;
-            }
-
-            currentNodeChain.Add(packageToProcess);
-            
-            foreach (ContentPackage dependency in dependencies)
-            {
-                if (!packages.ContainsKey(dependency))
-                {
-                    // search to see if it's enabled
-                    if (!ContentPackageManager.EnabledPackages.All.Contains(dependency))
-                    {
-                        ModUtils.Logging.PrintError($"Warning: the ContentPackage of {packageToProcess.Name} requires the Dependency {dependency.Name} but this package wasn't found in the enabled mods list!");
-                    }
-                    continue;
-                }
-
-                var ret = ProcessPackage(dependency, packages[dependency]);
-
-                switch (ret)
-                {
-                    case PackageProcRet.Cyclic:
-                        // TODO: Continue.
-                }
-            }
-
-            currentNodeChain.Remove(packageToProcess);
-            completedPackages.Add(packageToProcess);
-            
-            return PackageProcRet.Completed;
-        }
-    
-        private enum PackageProcRet : byte
-        {
-            Cyclic = 0,
-            AlreadyCompleted,
-            Completed,
-            BadPackage
-        }    
-    }
-    
-    
 
     #endregion
 }
